Update README.md
This is a Programming Logic project requested by the university. The task was to create an algorithm that involves classes, objects, and inheritance. The choice to focus on COVID-19 was made due to the timing of the project. The goal was to investigate the status of COVID-19 vaccination in the city of Brasília, as well as the manufacturing laboratory of the vaccine. The project was conducted on the Azure Lab Virtual Machine provided by the university.